rpma: common: separate common_td_port()
