+++ engines/librpma_common.h
@@ -1,0 +1,26 @@
+/*
+ * librpma_common: librpma and librpma_gpspm engine's common header
+ *
+ * Copyright 2021, Intel Corporation
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License,
+ * version 2 as published by the Free Software Foundation..
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ */
+
+#ifndef LIBRPMA_COMMON_H
+#define LIBRPMA_COMMON_H 1
+
+#include "../fio.h"
+
+#define LIBRPMA_COMMON_PORT_STR_LEN_MAX 12
+
+int librpma_common_td_port(const char *port_base_str, struct thread_data *td,
+	char *port_out);
+
+#endif /* LIBRPMA_COMMON_H */
